+++ client/commands/mod.rs
@@ -1,5 +1,4 @@
 use anyhow::Result;
-use async_std::net::TcpStream;
 
 use pueue::message::Message;
 use pueue::protocol::*;
@@ -11,7 +10,7 @@ pub mod restart;
 
 // This is a helper function for easy retrieval of the current daemon state.
 // The current daemon state is often needed in more complex commands.
-pub async fn get_state(socket: &mut TcpStream) -> Result<State> {
+pub async fn get_state(socket: &mut Socket) -> Result<State> {
     // Create the message payload and send it to the daemon.
     send_message(Message::Status, socket).await?;