+++ daemon/socket.rs
@@ -1,7 +1,6 @@
 use std::sync::mpsc::Sender;
 
 use anyhow::{bail, Result};
-use async_std::net::{TcpListener, TcpStream};
 use async_std::task;
 use log::{debug, info, warn};
 
@@ -16,19 +15,33 @@ use crate::streaming::handle_follow;
 /// Poll the unix listener and accept new incoming connections.
 /// Create a new future to handle the message and spawn it.
 pub async fn accept_incoming(sender: Sender<Message>, state: SharedState, opt: Opt) -> Result<()> {
-    // Commandline argument overwrites the configuration files values for port.
-    let port = if let Some(port) = opt.port.clone() {
-        port
-    } else {
+    let (unix_socket_path, port) = {
         let state = state.lock().unwrap();
-        state.settings.daemon.port.clone()
+        let shared = &state.settings.shared;
+
+        // Return the unix socket path, if we're supposed to use it.
+        if shared.use_unix_socket {
+            (Some(shared.unix_socket_path.clone()), None)
+        } else {
+            // Otherwise use tcp sockets and a given port
+            // Commandline argument overwrites the configuration files values for port.
+            let port = if let Some(port) = opt.port.clone() {
+                port
+            } else {
+                shared.port.clone()
+            };
+
+            (None, Some(port))
+        }
     };
-    let address = format!("127.0.0.1:{}", port);
-    let listener = TcpListener::bind(address).await?;
+
+    let listener = get_listener(unix_socket_path, port).await?;
 
     loop {
         // Poll if we have a new incoming connection.
-        let (socket, _) = listener.accept().await?;
+        let socket = listener.accept().await?;
+
+        // Start a new task for the request
         let sender_clone = sender.clone();
         let state_clone = state.clone();
         task::spawn(async move {
@@ -41,7 +54,7 @@ pub async fn accept_incoming(sender: Sender<Message>, state: SharedState, opt: O
 /// In case we received an instruction, handle it and create a response future.
 /// The response future is added to unix_responses and handled in a separate function.
 async fn handle_incoming(
-    mut socket: TcpStream,
+    mut socket: Socket,
     sender: Sender<Message>,
     state: SharedState,
 ) -> Result<()> {
@@ -59,7 +72,7 @@ async fn handle_incoming(
     // Return immediately, if we got a wrong secret from the client.
     {
         let state = state.lock().unwrap();
-        if secret != state.settings.daemon.secret {
+        if secret != state.settings.shared.secret {
             warn!("Received invalid secret: {}", secret);
             bail!("Received invalid secret");
         }
@@ -69,7 +82,7 @@ async fn handle_incoming(
     // locking the state in the streaming loop.
     let pueue_directory = {
         let state = state.lock().unwrap();
-        state.settings.daemon.pueue_directory.clone()
+        state.settings.shared.pueue_directory.clone()
     };
 
     loop {
@@ -81,11 +94,6 @@ async fn handle_incoming(
             // The client requested the output of a task.
             // Since we allow streaming, this needs to be handled seperately.
             handle_follow(&pueue_directory, &mut socket, &state, message).await?
-        } else if let Message::DaemonShutdown = message {
-            // Simply shut down the daemon right after sending a success response.
-            let response = create_success_message("Daemon is shutting down");
-            send_message(response, &mut socket).await?;
-            std::process::exit(0);
         } else {
             // Process a normal message.
             handle_message(message, &sender, &state)