+++ shared/protocol.rs
@@ -1,15 +1,17 @@
-use ::anyhow::{Context, Result};
-use ::async_std::net::TcpStream;
-use ::async_std::prelude::*;
-use ::byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
-use ::log::debug;
-use ::std::io::Cursor;
+use std::io::Cursor;
+
+use anyhow::{Context, Result};
+use async_std::prelude::*;
+use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
+use log::debug;
 
 use crate::message::*;
 
+pub use crate::platform::socket::*;
+
 /// Convenience wrapper around send_bytes.
 /// Deserialize a message and feed the bytes into send_bytes.
-pub async fn send_message(message: Message, socket: &mut TcpStream) -> Result<()> {
+pub async fn send_message(message: Message, socket: &mut Socket) -> Result<()> {
     debug!("Sending message: {:?}", message);
     // Prepare command for transfer and determine message byte size
     let payload = bincode::serialize(&message).expect("Failed to serialize message.");
@@ -19,7 +21,7 @@ pub async fn send_message(message: Message, socket: &mut TcpStream) -> Result<()
 
 /// Send a Vec of bytes. Before the actual bytes are send, the size of the message
 /// is transmitted in an header of fixed size (u64).
-pub async fn send_bytes(payload: Vec<u8>, socket: &mut TcpStream) -> Result<()> {
+pub async fn send_bytes(payload: Vec<u8>, socket: &mut Socket) -> Result<()> {
     let message_size = payload.len() as u64;
 
     let mut header = vec![];
@@ -39,7 +41,7 @@ pub async fn send_bytes(payload: Vec<u8>, socket: &mut TcpStream) -> Result<()> 
 
 /// Receive a byte stream depending on a given header.
 /// This is the basic protocol beneath all pueue communication.
-pub async fn receive_bytes(socket: &mut TcpStream) -> Result<Vec<u8>> {
+pub async fn receive_bytes(socket: &mut Socket) -> Result<Vec<u8>> {
     // Receive the header with the overall message size
     let mut header = vec![0; 8];
     socket.read(&mut header).await?;
@@ -75,7 +77,7 @@ pub async fn receive_bytes(socket: &mut TcpStream) -> Result<Vec<u8>> {
 }
 
 /// Convenience wrapper that receives a message and converts it into a Message.
-pub async fn receive_message(socket: &mut TcpStream) -> Result<Message> {
+pub async fn receive_message(socket: &mut Socket) -> Result<Message> {
     let payload_bytes = receive_bytes(socket).await?;
 
     // Deserialize the message.