+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.3]
+## [0.8.0]
+
+This version adds breaking changes:
+
+- The configuration file structure has been changed. There's now a `shared` section.
+- The configuration files have been moved to a dedicated `pueue` subdirectory.
+
+### Added
+
+- Unix socket support
 
 ### Fixed